SafeMode.ps1 v1
An exploratory tool for analyzing the service and driver dependencies of safeboot/safe mode. 

The script will export running driver and service information in each state then compared. This can be used to create a simlar operating state while booted normally, useful for troubleshooting or to reduce the time to create a minimal build/installation using safemode  configuration state as the baseline. The running system could be refined further with additional research. Disabling specific services or drivers can impact performance or normal operation, make incremental changes and test the impacts.