Videocard: remove advice for nv slider
The slider does not change many settings, but for the settings it does change,
it would be preferable to explicitly address each in the document, and what that setting does.
More importantly, the suggestion here will globally disable VSync, which
WILL disrupt video playback and WILL negatively affect gsync.

There's nothing to gain except potential confusing by suggesting this.